Add generic Jest padding rule
This padding rule is based on eslint/padding-line-between-statements.

`src/rules/padding.ts` reconfigures the eslint rule to look for Jest
tokens and add padding lines. There are several bits of functionality
that were not needed from the original and they have been removed. The
file is also converted to TypeScript to match this project's preferred
language.