Reproduce current describe and test rules using new generic rule
Introduces a `makeRule` function to take a set of padding configs, copy
the generic rule, configure it with the padding configs, and return it
as a new rule.

`makeRule` is now used in `src/index.ts` to generate the rules from a
config instead of defining rules in their own modules and importing them.

Minor doc and spec changes, as well.