Remove .nvmrc
The `engines` entry in `package.json` should be enough for this project.
`.nvmrc` doesn't give us as much flexibility in specifying a range of
acceptable node versions (by design) and we don't want to mandate an
arbitrary version for contributors using `nvm`.

`.nvmrc` does provide some aliases and wildcards, like `node` for latest
or `lts/*` for latest LTS, but it's still not quite what we want.

I've added `.nvmrc` to `.gitignore` so if a contributor would like to
add an `.nvmrc` for their own specific needs then they are welcome to do so
for their local dev.

fixes #97